Refactor: 파일 업로드 컴포넌트 분리
