+++ src/pages/WritePost.tsx
@@ -1,7 +1,13 @@
-import Write from './Write'
+import Write from '@src/components/Write/Write'
+import { useEffect } from 'react'
 
 const WritePost = () => {
-  console.log('분리예정')
+  // api호출해 state에 담고 form에 보내기
+
+  useEffect(() => {
+    console.log('write')
+  }, [])
+
   return <Write />
 }
